# lib/api-zod/src/generated/types/shopStaffCustomer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopUser } from './shopUser';
import type { ShopVehicle } from './shopVehicle';

export type ShopStaffCustomer = ShopUser & {
  vehicles: ShopVehicle[];
};