# lib/api-zod/src/generated/types/shopWorkOrderCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopWorkOrderCreateStatus } from './shopWorkOrderCreateStatus';

export interface ShopWorkOrderCreate {
  customerId: string;
  vehicleId: string;
  /** @minLength 1 */
  service: string;
  status?: ShopWorkOrderCreateStatus;
  /**
     * @minimum 0
     * @maximum 1
     */
  progress?: number;
  eta: string;
  technician: string;
  note: string;
  estimate: string;
  approved?: boolean;
}