# lib/api-zod/src/generated/types/shopWorkOrderCustomer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShopWorkOrderCustomer {
  id: string;
  name: string;
  email: string;
}