# lib/api-zod/src/generated/types/shopWorkOrderUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopWorkOrderUpdateStatus } from './shopWorkOrderUpdateStatus';

export interface ShopWorkOrderUpdate {
  service?: string;
  status?: ShopWorkOrderUpdateStatus;
  /**
     * @minimum 0
     * @maximum 1
     */
  progress?: number;
  eta?: string;
  technician?: string;
  note?: string;
  estimate?: string;
  approved?: boolean;
}